change bcc_chance from parseInt to parseFloat

+++ src/services/MessagingCampaignAPI.js
@@ -24,7 +24,7 @@ class MessagingCampaignAPI {
          * @type {AxiosPromise}
          */
         data.limit = Number.parseInt(data.limit);
-        data.bcc_chance = Number.parseInt(data.bcc_chance);
+        data.bcc_chance = Number.parseFloat(data.bcc_chance);
 
         const apiCompletionPromise = request({
             method: 'post',
